requirements.txt
absl-py==0.8.1
certifi==2019.9.11
future==0.18.1
grpcio==1.24.3
imageio==2.6.1
Markdown==3.1.1
numpy==1.17.3
pandas==0.25.2
Pillow==6.2.1
protobuf==3.10.0
python-dateutil==2.8.0
pytorch-lightning==0.5.2.1
pytz==2019.3
PyYAML==5.1.2
six==1.12.0
tb-nightly==1.15.0a20190708
test-tube==0.7.3
torch==1.3.0.post2
torchvision==0.4.1.post2
tqdm==4.36.1
Werkzeug==0.16.0
